+++ pkg/addons/addons.go
@@ -19,8 +19,8 @@ import (
 	"k8s.io/apimachinery/pkg/util/wait"
 
 	"github.com/openshift/openshift-azure/pkg/api"
-	"github.com/openshift/openshift-azure/pkg/jsonpath"
 	"github.com/openshift/openshift-azure/pkg/util/azureclient"
+	"github.com/openshift/openshift-azure/pkg/util/jsonpath"
 )
 
 const ownedBySyncPodLabelKey = "azure.openshift.io/owned-by-sync-pod"
