+++ pkg/addons/default.go
@@ -3,7 +3,7 @@ package addons
 import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 
-	"github.com/openshift/openshift-azure/pkg/jsonpath"
+	"github.com/openshift/openshift-azure/pkg/util/jsonpath"
 )
 
 func defaultContainerSpec(obj map[string]interface{}) {
