+++ pkg/arm/resources_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/openshift/openshift-azure/pkg/api"
-	"github.com/openshift/openshift-azure/pkg/jsonpath"
+	"github.com/openshift/openshift-azure/pkg/util/jsonpath"
 )
 
 func TestFixupDepends(t *testing.T) {
