+++ pkg/config/util.go
@@ -7,7 +7,7 @@ import (
 
 	"k8s.io/client-go/tools/clientcmd/api/v1"
 
-	"github.com/openshift/openshift-azure/pkg/tls"
+	"github.com/openshift/openshift-azure/pkg/util/tls"
 )
 
 func makeKubeConfig(clientKey *rsa.PrivateKey, clientCert, caCert *x509.Certificate, endpoint, username, namespace string) (*v1.Config, error) {
